window-actor/x11: Do not add shadows to undecorated shaped windows
They are extremely special (xeyes, xsnow, ...) and it's impossible
for us to second-guess the intended behavior with regard to shadows.

We did in fact not add a shadow in that case until commit b975676c
switched shadows from "default off" to "default on".

The motivation for that change were client-side (but not shaped) windows
like Chrome, and applying it also to shaped windows was arguably too
broad; at least it broke a particular OpenJDK compliance test for
shaped windows.

Given that XShape is increasingly niche, erring on the conservative
side and not add a shadow is unlikely to affect any apps/clients
where the shadow is genuinely missing.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4269>