inputcapture/session: Ensure we recreate devices after Disable/Enable
Disable/Enable may be called multiple times as the input capture client
is ready or not. Previously we unref'd the devices and the seat so
on Enable we no longer had those devices and events were captured but
discarded due to session->eis_pointer being NULL.

Fix this by re-creating the eis devices on Enable. The first time Enable
is called we probably don't have a seat yet so we use that for gating.

Also use clear_eis_pointer() instead of just the unref to ensure we
remove the device as seen by the client. The device may otherwise still
have a ref somewhere and thus may not be removed until later, causing
potential unexpected states.

Closes: https://gitlab.gnome.org/GNOME/mutter/-/issues/3908
Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4273>