frames/window-tracker: Handle override-redirect attribute change.
If a client creates windows with override-redirect attribute, but later
changes it to allow its windows to be managed, the window tracker will
still not be tracking _MUTTER_NEEDS_FRAME property changes and any
decoration change will be broken.

This reproduces with Wine, if a window is first created as a popup, to
later change to a framed window. Wine always creates its windows with
override-redirect, and only later decides to let them be managed.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4287>