onscreen/native: Peek the correct frame for power save page flips
The superseded frame is always the oldest frame so it's the head of
the queue, not the tail. Aside from anything else, the call to
`meta_onscreen_native_notify_frame_complete` that follows *does* use
the head and not the tail.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4310>