wayland/dma-buf: Use opaque format substitutes for scanout
Many display engines, especially older or simpler ones, only support
opaque formats. Many clients, however, require formats with alpha for
decorations and can't easily change formats on demand in situations
when their windows are fully opaque, e.g. fullscreen or tiled modes.
Instead they use opaque regions - a feature present both in Wayland
and EWMH - to let compositors know about partial or full opaqueness.

In cases of full opaqueness we can make use of this knowledge and
replace the alpha format with its opaque variant when handing buffers
to KMS for scanout. I.e. we "lie" to KMS about the format, because we
know that the visual result is equal to the one we'd get if the display
engine supported the alpha format.

As a result we hit direct scanout paths for common use-cases like
fullscreen video playback much more often on affected devices, reducing
resource usage.

This "trick" has been used by other compositors like Weston for many
years and appears to work with all upstream modifiers. In theory there
could be modifiers for which this assumption doesn't hold, however it
seems unlikely enought to not block on it for now.

In order to make clients (Mesa) switch to scanout modifiers/flags/etc.
accordingly, we also need to add the actually not supported alpha
formats in scanout tranches.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4320>