clutter/box-layout: Update actor request-mode when changing orientation
ClutterBoxLayout is changing the request-mode of the associated actor
depending on the orientation of the box (see the set_container() vfunc
implementation in ClutterBoxLayout). Apparently it was forgotten to update
the request-mode when setting box orientation via
clutter_box_layout_set_orientation().

This is a fairly substantial bug, as it means that children of vertical
BoxLayouts often get allocated a smaller height than their natural height,
here's a gnome-shell/St bug this triggers:

1) ClutterBoxLayout defaults to HORIZONTAL orientation, so when attaching
the box layout to a "fresh" actor, the actors request-mode is changed to
WIDTH_FOR_HEIGHT (away from the default HEIGHT_FOR_WIDTH).

2) Now StBoxLayout in gnome-shell has a "vertical" property, and no matter
when that property is set (even at construct time), the ClutterBoxLayout
is already attached to the actor, so request-mode will always be
WIDTH_FOR_HEIGHT for vertical StBoxLayouts.

3) WIDTH_FOR_HEIGHT request-mode means that
clutter_actor_adjust_allocation() is passing -1 as for_width to
get_preferred_height(), and that has impact on the reported natural height
of eg. labels (natural height will be just a single line).

4) In the end, an StLabel child of a vertical StBoxLayout will end up
getting a smaller allocation than it should, it'll allocate just a single
line instead of all lines of the label.

Fix the bug by adding the missing call to clutter_actor_set_request_mode()
in clutter_box_layout_set_orientation(). To reproduce, create an StLabel
with line_wrap=true and a vertical StBoxLayout that has a width small enough
so the label has to wrap. The label will not wrap, but instead show a single
line and then ellipsize. With this patch, it will wrap and show the full
label.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4339>
(cherry picked from commit 43ab8a774f105578e890f720bb02b798b2e0bd01)