constraints: Flip Wayland popups even if they won't full fit afterwards
We flip Wayland popups around the x/y axis if they don't fit into the
work-area otherwise. We currently only do this in case the popup fully fits
within the work-area after flipping.

Quite often popups have the ability to resize or slide though, so not all is
lost if the popup doesn't fit the work-area.

If the area of the work-area that a popup can occupy is higher when flipped
than without flipping, it still makes sense to flip the popup. It will only
work when RESIZE or SLIDE is allowed for the popup, but the same restriction
applies for the non-flipped case, so arguably we're getting the better
result by flipping.

When doing this in conjunction with SLIDE, it means that we flip to the side
where the intersection with the work-area is larger at first, and then slide
the popup to fully fit. In practice this means that the amount we have to
slide the popup by is smaller than without flipping, so it still makes sense
to flip.

When doing this	in conjunction with RESIZE, it means that less size will
have to be "cut away" from the popup, and a larger area of the popup will
end up on screen. Eg. for larger-than-monitor context-menus close to the
bottom screen edge, this means we still flip them, rather than leaving them
unflipped and super small.

FIXME: I'm not 100% sure about this, maybe we should check whether the
flipped+intersected rect still has the correct anchor. But if it doesn't,
that probably means the anchorpoint is outside the work-area anyway, so
without flipping we wouldn't show the popup at all (which might actually be
more correct?).

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4350>