test/kvm: Patch out c23 incompatibilities
The kernel defines a couple of symbols that conflict with c23,
which is the default for recent versions of gcc.

Patch the kernel to not redefine those symbols when using
a recent enough C standard to fix compilation with newer
gcc versions.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4349>