check-style: Double-check suggested changes
check-style.py was previously raising false positives, complaining about
style issues that didn't exist and then making bogus suggestions.
But `uncrustify` wasn't to blame, we were using it wrong.

In order to keep suggestions confined to just the chunks that changed
recently, `check-style.py` feeds those isolated chunks of code into
`uncrustify`. But that was the problem - we weren't giving `uncrustify`
enough context to make a good judgement on the correct alignment and
indentation.

But we don't want to style check the entire file every time a change is
made. Doing so will raise too many style complaints about legacy code.
So now we take the best of both worlds:

1. Run a first pass over just the code that changed recently and flag
   if any of it *might* need fixing.

2. Run a second pass over the entire file that *might* need fixing so that
   `uncrustify` can get the full context and eliminate any bad ideas it got
   in the first pass.

This has the benefit of eliminating bogus suggestions in files that have
perfect formatting when the whole file context is taken into account. The
drawback is that any files that need fixing outside of the current context
will still be subject to bogus suggestions. But this means you only need
to fix the whole file context once, and then it won't be subject to any
bogus suggestions in future.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4356>