cursor-sprite/xcursor: Add a cache for the XCursor images
The cache lives on the MetaCursorTracker and thus is shared between all
CursorSpriteXcursors and avoids constantly having to hit the disk when
the cursor shape changes.

We still do a lot of work to get the sprite from the theme into a KMS
plane or drawn into the onscreen, but avoiding disk IO is a first, good
step.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4359>