kms/impl-device/simple: Add support for missing KMS properties
While we do not intend to support HDR on the simple device impl, being
able to get to the default color mode requires support for those
properties. Otherwise, if the monitor is already in HDR mode when mutter
starts, we would be stuck there.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4357>