workspace: Announce workspace changes to screen reader
We currently play a sound effect on workspace switches (if supported
and the needed sounds are provided by the theme).

For visually impaired users, it would also be helpful if the screen
reader announced workspaces switches as well, so do that.

Closes: https://gitlab.gnome.org/GNOME/gnome-shell/-/issues/8301
Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4385>