backend: Make keyboard map setting async
On the native backend the keymap is effectively applied asynchronously
since applying happens cross threads. Some events coming from the input
thread also assumes there is a compatible keyboard map associated with
it, which might not be the case between the time the set_keymap() call
is made, and the keymap was actually applied in all places.

Address this by not setting the keymap in the main thread until the
input thread has signalled it have done so itself. This should allow us
to use the keymap in the main thread on events from the input thread
more reliably.

Also make sure to emit modifier state after changing the modifier
applies to the to the main thread new current map, and that the async
return source has the same priority as the input method callbacks, so
the order of events are preserved.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4394>