backend: Make function setting layout group async
It effectively already was, but make it even more so, so that the layout
group doesn't get updated on the main thread until the input thread is
done with the old layout group.

Also make sure to return set-keymap task before updating keymap state.
Returning the task will emit a keymap-changed signal on the main thread,
and updating the keymap state may emit a state-changed signal on the
keymap on the main thread. Make sure the keymap-changed signal is
emitted before the state-changed signal.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4394>