wayland: Avoid updating cursor on X11-driven drag-and-drop
In X11 we have a special grab that lets the X11 drag source be in
charge of driving the DnD operation. Between all those responsibilities,
one is updating the pointer cursor to reflect DnD feedback, so we
shouldn't do it in the compositor side in its behalf.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4390>