clutter: Trigger repick after all transitions are complete
The ClutterAnimatable::set_final_state() method is a bit misleading
since it does run for every frame. Since we do not have the mechanisms
to repick while an animation is ongoing other than enter/leave events,
truly postpone it until the animation is done.

Ideally, transitions should update the pointer position while they
are running, and that should trigger motion events to the right place,
and compositor animations could opt out of that on a per-effect basis
by making the actor insensitive, but we're not there yet.

Closes: https://gitlab.gnome.org/GNOME/mutter/-/issues/4052
Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4392>