pointer: Register implicit grab without racing against the release event
Normally on mouse button press and release we handle updating the button_count,
sending the event to clients, and registering the implicit grab serial all
synchronously, so there is no race.

However, if there is an active grab (such as an open popup) we delay sending
the event to clients and registering the implicit grab serial asynchronously
(with _clutter_stage_queue_event).

What can happen then is that there was a button release event already queued
which we start processing immediatly and update the button_count back to 0
before the clutter stage event gets processed (clutter_frame_clock_dispatch),
which would find no buttons pressed and thus does not register the implicit
grab serial.

Avoid the race by counting the pressed buttons at the moment of the event we
are dispatching.

Closes: https://gitlab.gnome.org/GNOME/mutter/-/issues/4055
Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4399>