backlight/ref-white: New backlight which adjusts the ref-white luminance
This backlight adjusts the reference white luminance level by a factor
to make the monitor brighter or darker.

This doesn't actually adjust the backlight which can be noticable on LCD
monitors. It also sacrificies the top range of the video signal which
can lead to banding artifacts.

Thus it is currently limited to HDR mode when a sysfs based backlight is
not found, because a lot of monitors completely disable "brightness"
control in HDR mode.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4014>