compositor-view/native: Limit direct scanout to RGB, YCbCr 709 limited
This makes sure we don't get wrong colors when a surface hits direct
scanout but need to be transformed with a different matrix than the one
that KMS uses.

We could be smarter about this and actually set the KMS state
approprietly when it's supported.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4408>