wayland/client: Create WaylandClient for every wl_client and adjust API
Before this change, MetaWaylandClient was only for clients that were
spawned, or otherwise created by the compositor. Clients connecting to
the normal wayland socket were not represented by it.

This change makes MetaWaylandClient generic, so that every wl_client has
a corresponding MetaWaylandClient.

It still does track how the client came about, via the kind.

It also adds a new capability system which can be used to assign
capabilities to clients that they usually don't, and that can be used in
wayland global filters. The service channel will make use of this.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4491>