place: Place window according to new size
Instead of using the current window size, apply placing algorithm using
the to be constrained size. This allows placing a window with a new size
before setting it as the current size.

Currently this makes practically no difference, since we always place
when showing a window, which we do when it already has a size, but it
will matter in the future where we want to avoid placing when showing
a window, in certain situations.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4413>