window: Wait with placing a non-floating window
A maximized or fullscreen window doesn't have a known floating size, so
placing it at this point, or saving a restored rect, isn't possible
without making up our own floating size. To not have to make up our own
size, we must postpone placing a window that was mapped non-floating until
it's floating. In the Wayland case, that means waiting until the clients
responds to a floating configure event with a committed floating
window state.

The test cases testing mapping as maximized is extended to test this case
as well. Another equivalent test case for fullscreen cases is added.

This also means some changes to how the monitor used for constraining or
placing is determined. For placing, the main monitor of the window is
used, unless the window is currently showing for the first time. This is
important due to the non-atomic nature of X11 window management, where a
window is configured step by step.

When constraining we always use the main monitor of the window if the
passed rectangle is not valid (e.g. an uninitialized unconstrained
window layout rectangle), and the window determined by the target window
layout rectangle if it is valid.

The saved rectangles are no longer saved on construct, so to maintain
some X11 specific behavior, the unmaximize and unfullscreen functions
was adapted to handle the case where we'd try to unmaximize/unfullscreen
to something where the client had already passed an unconstrained
rectangle.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4413>