window/wayland: Use last sent configuration to track resize changes
Instead of keeping around a few fields containing "last sent size"
(which isn't actually the last sent size), keep the last sent
configuration around.

Primarily it aims makes the MetaWindowWayland::move_resize_internal()
more easy to follow, with the requirements for needing to send new
configure events clearer, while making it possible to directly compare
previously sent configurations with new ones. What this allows is
constraining windows up-front in situations where it was previously
hard, e.g. acknowledged maximized configurations when the constraints
had changed. By always reconstraining acked window configurations that
are directly affected by constraints, e.g. maximized/fullscreen, we can
avoid a work around that queued extra move/resize's to "catch up" to
changed constraints. To avoid feedback loops where buggy client doesn't
respect fixed configurations, sending new configurations that are
equivalent to the last sent configuration is avoided.

When creating a new configuration, we didn't set the position in the
configuration if the position didn't changed, to avoid unexpected window
jumping. This was for the same reason done for surface state changes
where we'd create a new configuration and drop the position metadata.

This would however mean we couldn't use the fact whether a configuration
has a position or not tied to it, which is e.g. always the case with
fixed state / non-floating windows. So to make this value a bit more
sensible for those situations, don't clear it when a window isn't
floating.

Closes: https://gitlab.gnome.org/GNOME/mutter/-/issues/1627
Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4413>