tests: Add interactive move tests
This test checks that configuration from state changes (in this case
'active' (or focused)) doesn't negatively affect the window position
after an interactive move.

To be specific, this ensures that window configurations don't
incorrectly contain stale positions when the position changed, but no
new configuration was sent to reflect that.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4413>