window: Don't update saved rect when unmaximizing
We'd set up a "desired rect" that in fact is a pointer to the saved
rect, and manipulate it according to some heuristics about
auto-maximizing. Let's changed that to not updating the saved rect,
leaving it intact for future use. The automaximize heuristics is still
in place, but should perhaps be only enabled with the automaximize
preference turned on.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4413>