monitor-config-store: Also reset policy when resetting
Resetting is used only for test case scenarious, where the intention is
to reset the config store to it's initial state. Resetting the policy
was only done when setting a custom config. The policy test case relies
on being able to reload the configuration with a clean slate, and
currently this happens as a side effect to a custom configuration being
used prior to the policy test case is being run. Make the test case able
to run on itself by also resetting the policy state in reset().

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4418>