core: Add setter/getter for a11y properties in MetaWindow
We at the moment save the object path relative to the toplevel in
the a11y bus. This is an implementation detail of the existing
framework, so the specific data that we have to store relative to
a toplevel is expected to change in the future.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4542>