backends: Allow creating uninhibitable idle watches
Some idle watches are really only interested in whether there is user
activity, but not if there are any inhibiting clients. An example for
this would be the break timer in gnome-shell which should not be
affected by inhibitors.

Related: https://gitlab.gnome.org/GNOME/gnome-shell/-/issues/8378
Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4422>