window: Always auto-maximize in idle callback
Prior to this commit, auto-maximizing Wayland windows was not working
due to placing happening before showing.

However, instead of automatically maximizing during constraining after
placing, do so in a idle callback. This works more reliably primarily
because a Wayland window is placed on its first floating commit, at
which point we should not constrain according to being maximized when
applying the commit, as we'd place the window according to the
maximization constraint, not the state the window was committed as.

In other words, this avoids placing a large floatig window at 0,0, but
instead according to the appropriate placing algorithm, and then
maximized. To avoid the intermediate floating state showing, postpone
showing the window until any auto-maximization has completed.

Closes: https://gitlab.gnome.org/GNOME/mutter/-/issues/4094
Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4426>