wayland: Change MetaWindowDrag event handling for MetaWaylandToplevelDrag
Currently, the event handling to perform window drag operations during a DnD/toplevel
drag operation happens across a few layers. We on one hand have the DnD part of event
handling operating over MetaWaylandInput, and owning the grab in effect. On the other
hand we have the MetaWaylandToplevelDrag trying to make MetaWindowDrag work by
borrowing the MetaWaylandInput grab.

In between, there are the other stacked MetaWaylandEventHandlers that DnD chooses (not)
to work with, and in result maybe consuming certain events so they don't have a
chance to be processed. Preventing these handlers to run does also prevent the event
handler set up by MetaWaylandToplevelDrag/MetaWindowDrag from doing its part.

So we ideally want a way to borrow the same ClutterGrab within MetaWaylandInput,
and MetaWaylandToplevelDrag happens to be an object in the wayland namespace managed
by the wayland seat, eligible for using MetaWaylandInput. Use this astonishing
coincidence to change the way events are processed by MetaWaylandDrag:

Set up a MetaWaylandEventHandler at MetaWaylandToplevelDrag, made to cooperate with
the MetaWaylandDataDevice event handler right next in the chain, and feed events to
MetaWindowDrag via a new meta_window_drag_process_event() private function, instead
of letting it hook itself to the ClutterGrab underneath.

This inverts the handling order between MetaWaylandDataDevice and MetaWindowDrag,
now MetaWindowDrag is ensured to handle its own events first. This should not be
a source of bugs since the effective actions are pretty much mutually exclusive.

This approach is still somewhat ad-hoc, the reliance on MetaWaylandEventHandlers
stacking being just right is there, but it goes across many less layers than hooking
up on a pre-existing ClutterGrab. Now the only dubious interaction is the manual
event feeding to MetaWindowDrag, which is private between src/compositor and
src/wayland.