monitor: Make instance persistent until monitor is unplugged
This will eventually allow for adding API that allows users to listen on
updates directly on the monitor.

A couple of MetaOutput API functions get their pre-conditions relaxed a
bit to account for this. Setting the monitor if it's already set is
allowed when it's the same, as this allows persistent outputs, which is
already the case for virtual monitors.

Getting the monitor allows returning NULL, so that outputs without any
monitor associated with it can get one created.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4430>