color-store: Ignore removed files in ICC profile directory
When we receive a 'changed' event on the ICC profile directory, there is
a risk the file is removed before we have a chance to query it. In cases
like this, just ignore the file.

This avoids flaky color management test failures such as

    libmutter-FATAL-WARNING: Failed to query file info on
    '...goutputstream-AKCS62': Error when getting information for file
    “...goutputstream-AKCS62”: No such file or directory

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4431>