backends/native: Reunite accounting of pointer coordinates
There was this old pair of seat_impl->pointer_x/y variables carrying
the pointer coordinates. This was already redundant with the state
stored in MetaInputDeviceNative, use the opportunity to get rid of
it, in favor of the new priv->pointer_state point.

The default initial pointer position at 16,16 has also been removed
as it's been long ineffective. The upper layers have a better clue
of an ideal initial position, so 0,0 is as good of a bad guess.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4458>