tests/monitor-abstraction: Flip the order of external/builtin testing
The test checks that disabling/enabling monitors doesn't recreate the
instances. Next, the logical monitor bits are going to get tested, and
to make that more convenient, switch the order, so we can test that
disabling the monitor on the second logical monitor means the first
logical monitor stays alive. Prepare for that by switching the order,
which is harmless for this test as is.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4434>