mtk: Add dispose bin helper
This is a tiny helper that lets a user add data and GDestroyNotify
entries to a dispose bin, and then dispose it.

The intended use case is to pass it as an argument to functions that
needs to have various things disposed, but not immediately, but rather
later, as is decided by the caller.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4434>