x11: Drop session restore
As per
https://gitlab.gnome.org/GNOME/gnome-session/-/commit/2146124a64c2c7dd6697cc28f2471369ce0d1792
the session restore mechanism has been disabled on systemd systems for 5
years. No one complained.

On top of that,
https://gitlab.gnome.org/GNOME/gnome-session/-/merge_requests/139
removed that code as it
was dead and as a preperation of getting rid of X11 session.

So be procative and remove the dead-code from Mutter as well.

Part-of: <https://gitlab.gnome.org/GNOME/mutter/-/merge_requests/4438>